chore: refactors section utils
