+++ packages/section/__tests__/unit/buildSliceData.test.ts
@@ -1,0 +1,43 @@
+import { buildSliceData } from "../../src/utils/buildSliceData";
+
+describe("buildSliceData", () => {
+  it("should add elementId and ignoreSeparator properties", () => {
+    const originalData = [
+      { id: "a", name: "LeadersSlice" },
+      { id: "b", name: "DailyUniversalRegister" },
+      { id: "c", name: "OtherSlice" },
+      { id: "d", name: "LeadersSlice" },
+      { id: "e", name: "DailyUniversalRegister" },
+      { id: "f", name: "OtherSlice" },
+      { id: "g", name: "OtherSlice" },
+      { id: "h", name: "LeadersSlice" },
+      { id: "i", name: "OtherSlice" },
+      { id: "j", name: "OtherSlice" },
+      { id: "k", name: "OtherSlice" },
+      { id: "l", name: "DailyUniversalRegister" },
+    ];
+    const newData = buildSliceData(originalData);
+
+    expect(newData).toMatchSnapshot();
+  });
+
+  it("should not mutate passed data", () => {
+    const originalData = [
+      { id: "a", name: "LeadersSlice" },
+      { id: "b", name: "DailyUniversalRegister" },
+      { id: "c", name: "OtherSlice" },
+      { id: "d", name: "LeadersSlice" },
+      { id: "e", name: "DailyUniversalRegister" },
+      { id: "f", name: "OtherSlice" },
+      { id: "g", name: "OtherSlice" },
+      { id: "h", name: "LeadersSlice" },
+      { id: "i", name: "OtherSlice" },
+      { id: "j", name: "OtherSlice" },
+      { id: "k", name: "OtherSlice" },
+      { id: "l", name: "DailyUniversalRegister" },
+    ];
+    const json = JSON.stringify(originalData);
+    buildSliceData(originalData);
+    expect(JSON.stringify(originalData)).toEqual(json);
+  });
+});
