+++ packages/section/__tests__/unit/isSupplementSection.test.ts
@@ -1,0 +1,11 @@
+import { isSupplementSection } from "../../src/utils";
+
+describe("isSupplementSection", () => {
+  it("should return true for a known supplement section", () => {
+    expect(isSupplementSection("Times2")).toBeTruthy;
+  });
+
+  it("should return false for a unknown supplement section", () => {
+    expect(isSupplementSection("foo")).toBeFalsy;
+  });
+});
