+++ packages/section/src/utils/insertSectionAd.ts
@@ -1,0 +1,14 @@
+export const insertSectionAd = (isTablet: boolean) => (slices: any[]) => {
+  const adSlotIndex = 3; // 0 based index
+
+  if (!isTablet || slices.length <= adSlotIndex) return slices;
+
+  return [
+    ...slices.slice(0, adSlotIndex),
+    {
+      name: "SectionAd",
+      slotName: "native-section-ad",
+    },
+    ...slices.slice(adSlotIndex),
+  ];
+};
